refactor: cnosolidate embedding_serviec
