+++ api/app/services/embeddings_service/__init__.py
@@ -1,0 +1,8 @@
+"""
+Embeddings service package for the legal-search-RAG application.
+Provides vector embedding generation functionality.
+"""
+
+from app.services.embeddings_service.embedding_service import EmbeddingService
+
+__all__ = ["EmbeddingService"]
